Add name-server API
The netdev_stdlib types
(https://forge.puppetlabs.com/puppetlabs/netdev_stdlib) have a resource
`name_server` that does not have a provider in the cisco puppet module.
The node utils need a library to change the name servers first, so this
commit adds the ability to query, add, and remove nameservers.

It does not yet have vrf support.