VNI refactor #1
This is step 1 of the refactor, there is still plenty to fix in here.

The main thrust of this was adding platform checks for the n7k vs n9k differences. I suspect this needs more tweaking.

I updated the yaml to use key-val and added platform checks there too.

Fixed the feature methods.

Added additional tests to minitest along with platform checks.