Add command line options to update-hooks
The previous behavior of this script displayed, via git diff, all
updates to hooks installed by the script.  This requires user
interaction to page through the diff.  This was especially cumbersome on
the initial run, when the full contents of all hooks were displayed,
requiring the user to <spacebar> through pages of text.

This commit adds simple command line option parsing and recognizes the
following options:

--help         displays usage text

-v, --verbose  verbose output, including diffs of updated files
               (the diffs are no longer displayed by default)

--dry-run      Do not make any updates (but print the same messages
               that are normally printed to notify the user of which
               files are to be updated)