Enhance getting-started documentation
Add a note to CONTRIBUTING.md to let the developer know that the git
hooks require additional tools available in the PATH (e.g., rubocop).

Edit README-develop-node-utils-APIs.md to make the getting-start portion
of the docs a little more explicit for developers who may be just
starting out with git and/or ruby.  Reorganized a few paragraphs to more
clearly separate the things that are one-time setup vs. things that are
likely to be done on each feature branch.  Added a paragraph about
installing the project's git hooks.