Error if keyword substitution yields empty result
When the command_reference config_get_token and config_set methods use
keyword-based parameter substitution, array elements that contain a
parameter not supplied by the caller will be omitted from the result
array.  This behavior is desired to support, e.g., CLI interaction
involving VRFs, where both the presence or absence of a VRF are
permissible and should result in CLI interaction either containing or
not containing a vrf statement.  However, this behavior makes it easy
for developer oversight or typo to result in an empty command line
(which is bad).

This commit updates command_reference so that if the result of
keyword-based parameter substitution is an empty array then an
ArgumentError will be raised.  It also adds new assertions to test this
behavior to the existing test_config_get_token_hash_substitution test
case, and updates the documentation and changelog.