platform.rb: rescue/ensure does not return last eval'd expression
This method should return a hash but is returning an array instead.
The reason is because the `ensure` is causing the `pkg_hsh` at the end to not eval - which results in the last eval'd item to be the `pkgs` array.

I changed the rescue/ensure to a rescue/end and limited it to the config_get. Now it returns a hash as expected.