platform.rb: fix cache overwrite bug
Probably a day one issue for structured output. The problem was when a data item was not found, the `inv.select!` call was overwriting the actual `cache_hash` with a partially empty table; i.e.

```
cache_hash['cli_show']['show inventory'] => {"TABLE_inv"=>{"ROW_inv"=>[]}}
```
This caused future `Platform` calls to wrongly assume there was valid data in the cache - the cache wasn't empty because of the `"Table_inv"` entry but the data it pointed to was empty.